+++ R/SCDB-package.R
@@ -7,3 +7,4 @@ NULL
 #' @import utils
 utils::globalVariables(".")
 utils::globalVariables("NULLIF")
+utils::globalVariables("CONCAT_WS")